c3

+++ ss-local.sh
@@ -10,15 +10,21 @@ if [ ! -f "$INSTALL_FLAG" ]; then
 fi
 
 if [ $ENABLE_HTTP = "yes" ]; then
-	dpkg -i /app/cron_3.0pl1-128ubuntu2_amd64.deb
-	dpkg -i /app/libpopt0_1.16-10_amd64.deb
-	dpkg -i /app/logrotate_3.8.7-2ubuntu2_amd64.deb
-	dpkg -i /app/privoxy_3.0.24-1_amd64.deb
-	rm /etc/init.d/privoxy
+	INSTALL_FLAG="/app/privoxy.installed"	
+	
+	if [ ! -f "$INSTALL_FLAG" ]; then
+		dpkg -i /app/cron_3.0pl1-128ubuntu2_amd64.deb
+		dpkg -i /app/libpopt0_1.16-10_amd64.deb
+		dpkg -i /app/logrotate_3.8.7-2ubuntu2_amd64.deb
+		dpkg -i /app/privoxy_3.0.24-1_amd64.deb
+		rm /etc/init.d/privoxy
 
-	PRIVOXY_CONF="/app/privoxy.conf"
-	echo listen-address 0.0.0.0:$HTTP_PORT > $PRIVOXY_CONF
-	echo forward-socks5 / localhost:$SS_LOCAL_PORT . >> $PRIVOXY_CONF
+		PRIVOXY_CONF="/app/privoxy.conf"
+		echo listen-address 0.0.0.0:$HTTP_PORT > $PRIVOXY_CONF
+		echo forward-socks5 / localhost:$SS_LOCAL_PORT . >> $PRIVOXY_CONF
+		
+		touch $INSTALL_FLAG
+	fi
 	
 	/usr/sbin/privoxy $PRIVOXY_CONF
 fi
